Prometheus Monitoring Changes for Production Environment
Various changes were needed so Prometheus would play nice with Nginx.

First of all, prometheus itself was made available at intranet.redbutte.utah/prometheus in the nginx.conf file by way of proxy_pass to the prometheus docker container.

To better differentiate urls in django, the base urls for django_prometheus were changed to be prefixed by 'monitoring/'.

The docker-compose.production.yml was changed to include django's collectstatic command before serving the web app with gunicorn since it was removed from the Dockerfile in the previous commit. The prometheus service was also expanded to include access to SSL certs, and specify the url where prometheus would be served from.

The base docker-compose.yml was modified to specify docker-compose version 3.2 specifically.

The prometheus.yml file was edited to accommodate the above changes, particularly in regard to SSL certs and the metrics export path being changed to monitoring/metrics. Prometheus seemed unable to grab the metrics from just the docker service name so an additional target was added to the 'django' job for the production environment.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>