Snapshot of Github Issue
This commit reflects the code at the time I created a github issue for Snipe-it requesting help on snipe-it errors.

Nginx Dockerfile was expanded to install snipe-it (including php7.3) into the container.

nginx.conf was renamed to intranet.conf and a new file named nginx.conf was created. intranet.conf manages the location blocks of the various web services and nginx.conf configures the nginx process. It was created primarily to change the user of the process from nginx to www-data.

startup.sh was added to emulate the functions of the startup.sh script from the Snipe-it repo. It was modified to remove the commands related to apache, which isn't used here.

supervisord.conf is similarly based on a config file from the Snipe-it repo. It is expanded in this code base to also control php-fpm and nginx.

www.conf was added to specify the proper user and group used by php-fpm.

Various changes to the docker-compose files to install snipe-it into the nginx container rather than giving it it's own container.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>